refactor:  upgrade to edc 0.0.1-milestone-7 (#100)
refactor: rebranding groupdId into org.eclipse.edc and upgrade to edc 0.0.1-milestone-7
