+++ build.gradle.kts
@@ -13,8 +13,8 @@ val downloadArtifact: Configuration by configurations.creating {
 val identityHubVersion: String by project;
 val registrationServiceVersion: String by project;
 dependencies {
-    downloadArtifact("org.eclipse.dataspaceconnector.identityhub:identity-hub-cli:${identityHubVersion}:all")
-    downloadArtifact("org.eclipse.dataspaceconnector.registrationservice:registration-service-cli:${registrationServiceVersion}:all")
+    downloadArtifact("org.eclipse.edc:identity-hub-cli:${identityHubVersion}:all")
+    downloadArtifact("org.eclipse.edc:registration-service-cli:${registrationServiceVersion}:all")
 }
 
 // task that downloads the RegSrv CLI and IH CLI
@@ -23,15 +23,16 @@ val getJars by tasks.registering(Copy::class) {
 
     from(downloadArtifact)
         // strip away the version string
-        .rename { s -> s.replace("-${identityHubVersion}", "")
-            .replace("-${registrationServiceVersion}", "")
-            .replace("-all", "")
+        .rename { s ->
+            s.replace("-${identityHubVersion}", "")
+                .replace("-${registrationServiceVersion}", "")
+                .replace("-all", "")
         }
     into(layout.projectDirectory.dir("system-tests/resources/cli-tools"))
 }
 
 // run the download jars task after the "jar" task
-tasks{
+tasks {
     jar {
         finalizedBy(getJars)
     }
