--- extensions/policies/src/main/java/org/eclipse/dataspaceconnector/mvd/SeedPoliciesExtension.java
@@ -1,72 +1,0 @@
-/*
- *  Copyright (c) 2022 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial implementation
- *
- */
-
-package org.eclipse.dataspaceconnector.mvd;
-
-import org.eclipse.dataspaceconnector.policy.model.Permission;
-import org.eclipse.dataspaceconnector.spi.contract.offer.ContractDefinitionService;
-import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
-import org.eclipse.dataspaceconnector.spi.policy.PolicyEngine;
-import org.eclipse.dataspaceconnector.spi.policy.RuleBindingRegistry;
-import org.eclipse.dataspaceconnector.spi.system.Inject;
-import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
-import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
-import org.eclipse.dataspaceconnector.spi.types.TypeManager;
-
-import static org.eclipse.dataspaceconnector.spi.policy.PolicyEngine.ALL_SCOPES;
-
-/**
- * Extension to initialize the policies.
- */
-public class SeedPoliciesExtension implements ServiceExtension {
-
-    private static final String ABS_SPATIAL_POSITION = "ids:absoluteSpatialPosition";
-
-    /**
-     * Registry that manages rule bindings to policy scopes.
-     */
-    @Inject
-    private RuleBindingRegistry ruleBindingRegistry;
-
-    /**
-     * Policy engine.
-     */
-    @Inject
-    private PolicyEngine policyEngine;
-
-    @Inject
-    private TypeManager typeManager;
-
-    @Inject
-    private Monitor monitor;
-
-    @Override
-    public String name() {
-        return "Seed policies.";
-    }
-
-    /**
-     * Initializes the extension by binding the policies to the rule binding registry.
-     *
-     * @param context service extension context.
-     */
-    @Override
-    public void initialize(ServiceExtensionContext context) {
-        ruleBindingRegistry.bind("USE", ALL_SCOPES);
-        ruleBindingRegistry.bind(ABS_SPATIAL_POSITION, ContractDefinitionService.CATALOGING_SCOPE);
-
-        policyEngine.registerFunction(ALL_SCOPES, Permission.class, ABS_SPATIAL_POSITION, new RegionConstraintFunction(typeManager.getMapper(), monitor));
-    }
-
-}
