+++ gradle.properties
@@ -1,9 +1,9 @@
-edcGroup=org.eclipse.dataspaceconnector
-edcVersion=0.0.1-milestone-6
-registrationServiceGroup=org.eclipse.dataspaceconnector.registrationservice
-registrationServiceVersion=0.0.1-milestone-6-SNAPSHOT
-identityHubGroup=org.eclipse.dataspaceconnector.identityhub
-identityHubVersion=0.0.1-milestone-6-SNAPSHOT
+edcGroup=org.eclipse.edc
+edcVersion=0.0.1-milestone-7
+registrationServiceGroup=org.eclipse.edc
+registrationServiceVersion=0.0.1-20221122-SNAPSHOT
+identityHubGroup=org.eclipse.edc
+identityHubVersion=0.0.1-20221118-SNAPSHOT
 gatlingVersion=3.7.5
 assertj=3.23.1
 jupiterVersion=5.8.2
