+++ launchers/registrationservice/build.gradle.kts
@@ -23,7 +23,7 @@ val edcGroup: String by project
 val identityHubVersion: String by project
 val identityHubGroup: String by project
 val registrationServiceGroup: String by project;
-val registrationServiceVersion : String by project;
+val registrationServiceVersion: String by project;
 
 dependencies {
     implementation("${registrationServiceGroup}:registration-service:${registrationServiceVersion}")
@@ -33,30 +33,31 @@ dependencies {
 
     implementation("${edcGroup}:identity-did-web:${edcVersion}")
     implementation("${edcGroup}:identity-did-core:${edcVersion}")
-    implementation("${edcGroup}:core-base:${edcVersion}")
-    runtimeOnly("${edcGroup}:core-boot:${edcVersion}")
+    implementation("${edcGroup}:connector-core:${edcVersion}")
+    runtimeOnly("${edcGroup}:boot:${edcVersion}")
     implementation("${edcGroup}:control-plane-core:${edcVersion}")
-    implementation("${edcGroup}:observability-api:${edcVersion}")
-    implementation("${edcGroup}:core-micrometer:${edcVersion}")
+    implementation("${edcGroup}:api-observability:${edcVersion}")
+    implementation("${edcGroup}:micrometer-core:${edcVersion}")
     runtimeOnly("${edcGroup}:jetty-micrometer:${edcVersion}")
     runtimeOnly("${edcGroup}:jersey-micrometer:${edcVersion}")
-    implementation("${edcGroup}:filesystem-configuration:${edcVersion}")
+    implementation("${edcGroup}:configuration-filesystem:${edcVersion}")
+    implementation("${identityHubGroup}:identity-hub-api:${identityHubVersion}")
     implementation("${identityHubGroup}:identity-hub-credentials-verifier:${identityHubVersion}")
 
     // JDK Logger
-    implementation("${edcGroup}:jdk-logger-monitor:${edcVersion}")
+    implementation("${edcGroup}:monitor-jdk-logger:${edcVersion}")
 
     // To use FileSystem vault e.g. -DuseFsVault="true".Only for non-production usages.
     val useFsVault: Boolean = System.getProperty("useFsVault", "false").toBoolean()
     if (useFsVault) {
-        implementation("${edcGroup}:filesystem-vault:${edcVersion}")
+        implementation("${edcGroup}:vault-filesystem:${edcVersion}")
     } else {
-        implementation("${edcGroup}:azure-vault:${edcVersion}")
+        implementation("${edcGroup}:vault-azure:${edcVersion}")
     }
 }
 
 application {
-    mainClass.set("org.eclipse.dataspaceconnector.boot.system.runtime.BaseRuntime")
+    mainClass.set("org.eclipse.edc.boot.system.runtime.BaseRuntime")
 }
 
 tasks.withType<com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar> {
