+++ launchers/registrationservice/logging.properties
@@ -1,7 +1,7 @@
-handlers = java.util.logging.ConsoleHandler
-.level = INFO
-java.util.logging.ConsoleHandler.level = ALL
-java.util.logging.ConsoleHandler.formatter = java.util.logging.SimpleFormatter
-java.util.logging.SimpleFormatter.format = %1$tF %1$tT %4$s : %5$s %n
-org.eclipse.dataspaceconnector.level = FINE
-org.eclipse.dataspaceconnector.handler = java.util.logging.ConsoleHandler
+handlers=java.util.logging.ConsoleHandler
+.level=INFO
+java.util.logging.ConsoleHandler.level=ALL
+java.util.logging.ConsoleHandler.formatter=java.util.logging.SimpleFormatter
+java.util.logging.SimpleFormatter.format=%1$tF %1$tT %4$s : %5$s %n
+org.eclipse.edc.level=FINE
+org.eclipse.edc.handler=java.util.logging.ConsoleHandler
