+++ system-tests/resources/vault/company1/company1-vault.properties
@@ -11,5 +11,4 @@
 #       Microsoft Corporation - initial API and implementation
 #
 #
-
 company1assets-key1=key1
