+++ system-tests/resources/vault/company2/company2-vault.properties
@@ -11,5 +11,4 @@
 #       Microsoft Corporation - initial API and implementation
 #
 #
-
 company2assets-key1=key2
