+++ system-tests/src/test/java/org/eclipse/edc/system/tests/utils/TransferRequestFactory.java
@@ -1,0 +1,20 @@
+/*
+ *  Copyright (c) 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial implementation
+ *
+ */
+
+package org.eclipse.edc.system.tests.utils;
+
+import java.util.function.Function;
+
+public interface TransferRequestFactory extends Function<TransferSimulationUtils.TransferInitiationData, String> {
+}
